Refactor: Renamed "Model" folder into "models".
