+++ app/src/main/java/com/example/bookworms/fragments/AddPostFragment.kt
@@ -7,7 +7,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
 import com.example.bookworms.R
-import com.example.bookworms.Model.entities.PostEntity
+import com.example.bookworms.models.entities.PostEntity
 import com.google.android.material.textfield.TextInputEditText
 
 /**
@@ -34,21 +34,10 @@ class AddPostFragment : Fragment() {
 
     }
 
-    private fun initDummyPost() {
-        dummyPost = PostEntity(
-            id = "1",
-            title = "Harry Potter and the Philosopher's Stone",
-            description = "Harry Potter has never even heard of Hogwarts when the letters start dropping on the doormat at number four, Privet Drive. Addressed in green ink on yellowish parchment with a purple seal, they are swiftly confiscated by his grisly aunt and uncle. ",
-        )
-    }
-
     private fun populateUiWithDummyPost(view: View) {
         val titleInput = view.findViewById<TextInputEditText>(R.id.post_title_input)
         val descriptionInput = view.findViewById<TextInputEditText>(R.id.post_description_input)
         val addPostButton = view.findViewById<View>(R.id.addPostButton)
-
-//        titleInput.setText(dummyPost.title)
-//        descriptionInput.setText(dummyPost.description)
         addPostButton.setOnClickListener {
             // Add post to database
             Toast.makeText(context, "Post added successfully", Toast.LENGTH_SHORT).show()
