+++ app/src/main/java/com/example/bookworms/models/ModelRoom/dao/UserDao.kt
@@ -1,4 +1,4 @@
-package com.example.bookworms.Model.ModelRoom.dao
+package com.example.bookworms.models.ModelRoom.dao
 
 import androidx.lifecycle.LiveData
 import androidx.room.Dao
@@ -7,7 +7,7 @@ import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import androidx.room.Update
-import com.example.bookworms.Model.entities.User
+import com.example.bookworms.models.entities.User
 
 
 @Dao
