+++ app/src/main/java/com/example/bookworms/models/entities/BookEntity.kt
@@ -1,4 +1,4 @@
-package com.example.bookworms.Model.entities
+package com.example.bookworms.models.entities
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
