+++ app/src/main/java/com/example/bookworms/models/entities/User.kt
@@ -1,4 +1,4 @@
-package com.example.bookworms.Model.entities
+package com.example.bookworms.models.entities
 
 import androidx.annotation.NonNull
 import androidx.room.Entity
@@ -9,7 +9,6 @@ import java.io.Serializable
 @Entity(tableName = "users")
 data class User(
     @PrimaryKey
-    @NonNull
     var uid: String,
 
     var phone: String,
