+++ app/src/main/java/com/example/bookworms/viewModels/UserViewModel.kt
@@ -1,7 +1,7 @@
 package com.example.bookworms.viewModels
 
 import androidx.lifecycle.ViewModel
-import com.example.bookworms.Model.entities.User
+import com.example.bookworms.models.entities.User
 import com.example.bookworms.services.JoinedUserModel
 
 class UserViewModel :ViewModel(){
